Add problem matcher to VSCode extension definition.
This change to package.json adds a problem matcher that supports parsing
error messages from the Odin compiler within VSCode when executed via
tasks in tasks.json. This allows compile errors to show up properly in
the Problems panel, though may not be necessary if/when ols supports
`textDocument/publishDiagnostics` in the future.